requirements.txt
python-telegram-bot==13.15
feedparser==6.0.11
googletrans==4.0.0-rc1
requests==2.31.0